cleaning up a few things, moving bsky.py tests to bsky_test.py
